factor particle system parameterization to separate structure to improve code structure and maintainability
